Format Tables for Output

Applies common formatting to tables output in Quarto, R Markdown, etc.

Usage

bhhi_format_table(
  table,
  full_width = TRUE,
  left_align_stub = TRUE,
  bottom_align_column_labels = TRUE
)

Arguments

table

A gt:gt table object

full_width

Table will fill the entire width of its container. Default is true.

left_align_stub

Text in table stub is left-aligned. Default is true.

bottom_align_column_labels

Text in column labels and spanners is bottom aligned vertically. Default is true.

Value

The formatted table object.

Details

Currently sets table to full width and left aligns text in the table stub.

Examples

if (FALSE) { # interactive()
mtcars |>
  dplyr::as_tibble(rownames = "car") |>
  gt::gt(rowname_col = "car") |>
  bhhi_format_table()
}